allwinner: Add aw_de2 support
This is a DRM driver for Allwinner DE2 (display engine 2).
For now it supports only the typical HDMI pipeline using MIXER1/TCON1.
Some resolutions are known to cause problem (720x576 for example), this is likely
an interlacing problem.
VI layers with multiple YUV framebuffer cause weird problem (tested with gstreamer
and the kms backend).
aw_de2_dw_hdmi needs to be split as most of it is just for the DesignWare/Synopsis
HDMI controller and only a small glue part is needed for Allwinner. And the interface
for it will likely to be a generic one for DRM encoder/bridge so ANX6345 (RGB to eDP)
could use it too.