drm: Put fbsd vt function to its own file
I'm not yet sure what is the best way to deal with framebuffer being
different between Linux and FreeBSD but for now put FreeBSD specific
function in its own file as it's easier to sync drm_fb_helper.c with
upstream.
It's also unclear who wrote thoses function so no licence for now.

Signed-off-by: Emmanuel Vadot <manu@freebsd.org>